Do not try to delete a log file that has already been deleted, (not sure why that happened).

+++ src/core/libs/core/utilities/logger/LogFile.php
@@ -297,8 +297,10 @@ EOD;
 					/** @var \Core\Filestore\Backends\FileLocal $file */
 					++$x;
 					if($x > $limit){
-						$file->delete();
-						++$deleted;
+						if($file->exists()){
+							$file->delete();
+							++$deleted;
+						}
 					}
 				}
